Also move locale.c and culture-info.h
